feat(app): move django admin config
